moves all page level components to ts
