+++ client/pages/testimonials.tsx
@@ -3,13 +3,12 @@ import { animated, useTransition } from 'react-spring';
 import { Card, CardBody, CardHeader } from 'grommet';
 import { isMobile, isTablet } from 'react-device-detect';
 import { gql } from '@apollo/client';
-import PropTypes from 'prop-types';
 
 import CardItem from '../components/CardItem';
 import CardWrapper from '../components/CardWrapper';
 import { CardGridStyles, CardPageStyles, CardLinkStyles } from '../components/styles/CardPageStyles';
 import SEO from '../components/SEO';
-import { testimonialsPageImage } from '../content/mainPageImages.json';
+import { testimonialsPageImage } from '../content/mainPageImages';
 import { client } from './_app';
 
 const GET_TESTIMONIALS_QUERY = gql`
@@ -32,7 +31,7 @@ const GET_TESTIMONIALS_QUERY = gql`
 const TestimonialsPage = ({ testimonials, metadata }) => {
     const [showText, setShowText] = useState(null);
 
-    const displayItemTextHandler = item => setShowText(item);
+    const displayItemTextHandler = (item) => setShowText(item);
 
     const transition = useTransition(showText, {
         from: { opacity: 0, transform: 'translate3d(0,-40px,0)' },
@@ -99,7 +98,7 @@ const TestimonialsPage = ({ testimonials, metadata }) => {
                                             </div>
                                         </Card>
                                     </animated.div>
-                                )
+                                ),
                         )}
                 </CardGridStyles>
             </CardPageStyles>
@@ -119,10 +118,5 @@ export async function getStaticProps() {
         },
     };
 }
-
-TestimonialsPage.propTypes = {
-    testimonials: PropTypes.array,
-    metadata: PropTypes.object,
-};
 
 export default TestimonialsPage;
